feat: move client to controlplane package
