+++ services/controlplane/client.go
@@ -1,4 +1,4 @@
-package client
+package controlplane
 
 import (
 	"bytes"
